+++ Runtime/Variables/ColorVariable.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace Codetox.Variables
+{
+    [CreateAssetMenu(menuName = Framework.MenuRoot.Variables.Color, fileName = nameof(ColorVariable), order = 30)]
+    public sealed class ColorVariable : Variable<Color>
+    {
+    }
+}